Improve/fix semantic version parsing and comparison.
Versions with pre-release suffixes are no longer excluded, they are less
than their associated release and greater than the previous release.
Wildcard ranges automatically include pre-releases at the start and
exclude them at the end.
Carret (^, same major) and tilde (~, same major+minor) ranges exclude
pre-releases at the end, but not at the start without lowering it with
an empty pre-release (trailing -), e.g. "~1.15-".
The upper bound has to be specified with a pre-release now to exclude any
builds for its release, e.g. "<1.15-" to exclude 1.15-alpha.1, otherwise
it will be allowed.
For maximum ease of use, the wildcard form ("1.14.x", "1.x") should be
chosen, alternatively the ^ or ~ form if the start range can't be .0.