Allow FabricTweaker to operate as primary or secondary tweaker
When operating as primary tweaker, it provides all launch arguments.
When operating as secondary tweaker, it only provides the class transformer and
mod loading functionality but not the launch arguments.

This distinction is necessary because two tweakers providing arguments will
invariably lead to arguments being passed multiple times which is not allowed
for some of them.