Move server entrypoint in 20w22a+ before `server.properties` is loaded.
This is to solve two issues with the current entrypoint:
`server.properties` is loaded before all mods
Mods cannot register their own `GeneratorType` on the dedicated server before the generator type is selected. Late init is possible for this but it increases the late-init hack count to 2 for a future generator type api.

Also there is an issue with biomes api where the nether's generator type is ALWAYS forced to be vanilla nether generator due to late init hack always using the default nether generator.

However this solution comes at the cost of mod init being done before the eula is checked, so the game may be terminated before the server is started. Moving the eula logic ahead of mod init is out of scope for this PR. Regarding vanilla, registries are all boostraped before `server.properties` is loaded anyways, so vanilla does not really optimize game loading here. Long term it would be nice if the eula check was the first thing done in the future if mojang ever reads this pr :)