Initialize mods' root path lazily.
This is a partial workaround for issue #152 where MC unconditionally
creates a file system for its Jar that conflicted with loader doing the
same earlier. A complete fix that also works if a mod happens to access
ModContainer.getRootPath for MC needs to prevent MC from creating the
already existing FileSystem and closing it.

Co-authored-by: shedaniel <daniel@shedaniel.me>