Improve resolution errors.
This change makes the solver also consider replacing mods that aren't invalid deps for other mods, but have unmet deps themselves.

Mod installation and replacement solutions now provide a proper upper bound and a better lower bound instead of just a somewhat bogus potential version.

Most removal solutions are now instructions to replacing a mod that is compatible with a set of requirements.