feat: add ability to customize fabric directories
Adds the ability to customize the "mods" and "config" directories via
the "fabric.dir.mods" and "fabric.dir.config" Java properties. This
can allow for different launcher profiles to have different mod
directories so users don't need to install third party launchers to fix
conflicting mods or config directories while keeping all of their
settings, screenshots, and other preferences the same.

In the future, it might be a good idea to add a way to customize this in
the fabric installer.