Add flag to skip cache on reading manifest
JarURLConnection contains internal caching of opened JARs. This causes the JVM to maintain a handle to any JAR file accessed this way.

Said handle remains open until the cache invalidates or the JVM terminates. In the case of a short-lived program such as Fabric Loader, it is usually JVM termination. This prevents us from deleting the temporary remapping files after reading it's manifest.

Since this cache is theoretically useful in the context of Log4jLogHandler, leave it up to the caller to decide whether or not it should be bypassed.